split up product requests into their own sub-packages and modules
